requirements.txt
openai
requests
pyyaml
python-dotenv
gitpython
PyGithub
python-gitlab
google-generativeai>=0.4.1
anthropic
openai>=1.0.0
